Create a calculator mockup with basic CSS
HTML file created with a some buttons to resemble a calculator and its
display. A basic styling has been applied to the buttons and body.
So far, the JS file has only been created with a basic console.log to
test the link between the HTML file and the script.

Next step is to create the basic functions and test them through the
console.