Add number buttons functionality
Created event listeners for the number buttons and for the C (delete
everything) button. Added functionality to the decimal point button
with a preventive method so only one decimal point can be added.

Note: cutting this commit message short because there was a power
outage and my PC is running on the No break power only.