Add functionality to operators buttons
Created a function that gets the current number and stores it in the
first position of an array.
This function is called every time an operator or the "=" buttons are
pressed. The array index is moved to the second position and the new
current number is stored there.
After each operation is done, the result is stored at the first position
of the array, this way the previews result is ready to be used for the
next operation.

DEL button removes the last digit of the current number, and if the user
tries to remove the last digit, the current number is changed to
cero "0".

An event listener for the "+/-" button multiplies the current number by
minus 1 (-1).

Fixed a bug where the pressing the "=" button before the user makes an
input set the first number to 0.

AC button works fine, deletes every value.

Added an easter egg hidden in plain sight ;)

Known bugs: operations where both numbers are floats give weird results
and pushes numbers outside the calculator "screen" area.
Committing before an attempt to fix this is made.